refactor: モジュール化
TCPコネクション部分をある程度レイヤー化する。
`ractor_cluster` を使うのもありだとは思うんだけど、
プロダクションレディじゃないらしい。
まあ自前で組んでも保証されてないのは同じじゃないかと思うけど、
とりあえず解決しないといけないのは通信交換なので、
後から差し替えやすいようにして、とりあえずTCPで実験をする。
