+++ kaisui-ractor/src/actors/mod.rs
@@ -1,0 +1,7 @@
+pub mod registry_actor;
+pub mod text_actor;
+pub mod transport_client_actor;
+
+pub use registry_actor::RegistryActor;
+pub use text_actor::TextActor;
+pub use transport_client_actor::TransportClientActor;
